Move rule-specific config into rules module
The core module should not contain rule-specific stuff.
