+++ scalafix-rules/src/main/scala/scalafix/internal/rule/ExplicitResultTypesConfig.scala
@@ -1,8 +1,9 @@
-package scalafix.internal.config
+package scalafix.internal.rule
 
 import metaconfig._
 import metaconfig.generic.Surface
 import metaconfig.annotation._
+import scalafix.internal.config._
 
 case class ExplicitResultTypesConfig(
     @Description("Enable/disable this rule for defs, vals or vars.")
